# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de gastos e informações financeiras",
        "title": "Contabius API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://vkunssec.vercel.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/vkunssec/contabius/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "${HOST}",
    "basePath": "/",
    "paths": {
...
        "/revenues": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para retornar as receitas",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Revenues"
                ],
                "summary": "Rota para retornar as receitas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "678079f6f5080a39a8eedc1e"
                        ],
                        "description": "Ids das receitas a serem retornadas",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Revenues"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para criar uma receita",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Revenues"
                ],
                "summary": "Rota para criar uma receita",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da receita",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RevenueRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
...
        "/revenues/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para atualizar uma receita",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Revenues"
                ],
                "summary": "Rota para atualizar uma receita",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da receita",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Revenues"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID da receita",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para logar um usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Rota para logar um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados do usuário",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        },
...
        "/user/update/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rota para atualizar um usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Rota para atualizar um usuário",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados do usuário",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserUpdateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/BadRequest"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "BadRequest": {
            "description": "Estrutura padrão de resposta de erro de requisição",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Erro ao processar a requisição"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "Currency": {
            "type": "string",
            "enum": [
                "BRL",
                "USD",
                "EUR"
            ],
            "x-enum-varnames": [
                "CurrencyBRL",
                "CurrencyUSD",
                "CurrencyEUR"
            ]
        },
...
        "Method": {
            "type": "string",
            "enum": [
                "credit",
                "debit",
                "pix",
                "boleto",
                "cash"
            ],
            "x-enum-varnames": [
                "MethodCredit",
                "MethodDebit",
                "MethodPix",
                "MethodBoleto",
                "MethodCash"
            ]
        },
        "MethodId": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "MethodCreditId",
                "MethodDebitId",
                "MethodPixId",
                "MethodBoletoId",
                "MethodCashId"
            ]
        },
        "Methods": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID do método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/MethodId"
                        }
                    ],
                    "example": 1
                },
                "method": {
                    "description": "Método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Method"
                        }
                    ],
                    "example": "credit"
                }
            }
        },
        "Money": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "default code BRL",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Currency"
                        }
                    ],
                    "example": "BRL"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                }
            }
        },
...
        "Response": {
            "description": "Estrutura padrão de resposta",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Dados da resposta"
                },
                "message": {
                    "description": "Mensagem descritiva",
                    "type": "string",
                    "example": "Operação realizada com sucesso"
                },
                "success": {
                    "description": "Indica se a operação foi bem sucedida",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "RevenueRequest": {
            "type": "object",
            "required": [
                "amount",
                "revenue"
            ],
            "properties": {
                "amount": {
                    "description": "Valor da receita",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Money"
                        }
                    ]
                },
                "method": {
                    "description": "Método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Methods"
                        }
                    ]
                },
                "revenue": {
                    "description": "Receita",
                    "type": "string",
                    "example": "salário"
                }
            }
        },
        "Revenues": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Valor da receita",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Money"
                        }
                    ]
                },
                "created_at": {
                    "description": "Data de criação da receita",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "id": {
                    "description": "ID da receita",
                    "type": "string",
                    "example": "678079f6f5080a39a8eedc1e"
                },
                "method": {
                    "description": "Método de pagamento",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Methods"
                        }
                    ]
                },
                "revenue": {
                    "description": "Receita",
                    "type": "string",
                    "example": "salário"
                },
                "updated_at": {
                    "description": "Data de atualização da receita",
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                }
            }
        },
        "UserLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email do usuário",
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "description": "Senha do usuário",
                    "type": "string",
                    "example": "password123"
                }
            }
        },
...
        "UserType": {
            "type": "string",
            "enum": [
                "admin",
                "user",
                "guest"
            ],
            "x-enum-varnames": [
                "UserTypeAdmin",
                "UserTypeUser",
                "UserTypeGuest"
            ]
        },
...
        "UserUpdateRequest": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "description": "Email do usuário",
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "description": "Nome do usuário",
                    "type": "string",
                    "example": "John Doe"
                },
                "type": {
                    "description": "Tipo de usuário",
                    "allOf": [
                        {
                            "$ref": "#/definitions/UserType"
                        }
                    ],
                    "example": "admin"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}